Library XML: should work
