+++ part_1/src/com/chernova/libraryXML/Library.java
@@ -2,7 +2,6 @@ package com.chernova.libraryXML;
 
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
-import org.w3c.dom.Node;
 
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
@@ -20,6 +19,7 @@ public class Library {
 
     public static LinkedList<Book> list = new LinkedList<>();
     public static int id = 0;
+
     //получить все книги (метод ВОЗВРАЩАЕТ список всех книг в библиотеке)
     public static LinkedList<Book> getAllBooks() {
         return list;
@@ -31,7 +31,6 @@ public class Library {
         book.setId(++id);
         book.setLocalDate(new Date());
         list.add(book);
-
     }
 
 
@@ -42,44 +41,43 @@ public class Library {
 
         System.out.println("Изменить название книги: ");
         System.out.println("(оставьте поле пустым, если не хотите его менять)");
-        Main.sc.nextLine();
-        String title = Main.sc.nextLine();
+        Menu.sc.nextLine();
+        String title = Menu.sc.nextLine();
         if (!title.trim().isEmpty()) {
             book.setTitle(title);
         }
 
         System.out.println("Изменить содержание книги: ");
         System.out.println("(оставьте поле пустым, если не хотите его менять)");
-        String text = Main.sc.nextLine();
+        String text = Menu.sc.nextLine();
         if (!text.trim().isEmpty()) {
             book.setText(text);
         }
 
         System.out.println("Изменить жанр книги: ");
-        System.out.println(Main.ANSI_YELLOW);
+        System.out.println(Menu.ANSI_YELLOW);
         Genre.printGenre();
-        System.out.println(Main.ANSI_RESET);
+        System.out.println(Menu.ANSI_RESET);
         boolean readyToSet = false;
         while (!readyToSet) {
             try {
                 readyToSet = true;
-                int genreAdd = Main.sc.nextInt();
+                int genreAdd = Menu.sc.nextInt();
                 book.setGenre(Genre.values()[genreAdd - 1]);
             } catch (InputMismatchException e) {
                 System.out.println("Введите номер жанра");
                 readyToSet = false;
-                Main.sc.next();
+                Menu.sc.next();
             } catch (ArrayIndexOutOfBoundsException e) {
                 System.out.println("Выберите жанр из списка");
                 readyToSet = false;
             }
         }
 
-
         System.out.println("Изменить ISBN в формате: х-хххх-хххх-х ");
         System.out.println("(оставьте поле пустым, если не хотите его менять)");
-        Main.sc.nextLine();
-        String inputISBN = Main.sc.nextLine();
+        Menu.sc.nextLine();
+        String inputISBN = Menu.sc.nextLine();
         if (!inputISBN.trim().isEmpty()) {
             boolean check = false;
             while (!check) {
@@ -87,16 +85,15 @@ public class Library {
                     check = true;
                 } else {
                     System.out.println("Введенный isbn  не соответствует формату. Попробуйте еще раз ");
-                    inputISBN = Main.sc.nextLine();
+                    inputISBN = Menu.sc.nextLine();
                 }
             }
+            book.setISBN(inputISBN);
         }
-        book.setISBN(inputISBN);
-
 
         System.out.println("Изменить дату публикации в формате: yyyy-MM-dd ");
         System.out.println("(оставьте поле пустым, если не хотите его менять)");
-        String publishDate = Main.sc.nextLine();
+        String publishDate = Menu.sc.nextLine();
         if (!publishDate.trim().isEmpty()) {
             boolean check = false;
             while (!check) {
@@ -104,14 +101,14 @@ public class Library {
                     check = true;
                 } else {
                     System.out.println("Введенная дата не соответствует формату. Попробуйте еще раз ");
-                    publishDate = Main.sc.nextLine();
+                    publishDate = Menu.sc.nextLine();
                 }
             }
+            book.setPublishDate(publishDate);
         }
-        book.setPublishDate(publishDate);
 
         book.setLocalDate(new Date()); //дата изменения книги
-        System.out.println(Main.ANSI_GREEN + "Данные изменены" + Main.ANSI_RESET);
+        System.out.println(Menu.ANSI_GREEN + "Данные изменены" + Menu.ANSI_RESET);
     }
 
 
@@ -126,111 +123,63 @@ public class Library {
             }
         }
         if (isDelete) {
-            System.out.println(Main.ANSI_GREEN + "Книга успешно удалена" + Main.ANSI_RESET);
+            System.out.println(Menu.ANSI_GREEN + "Книга успешно удалена" + Menu.ANSI_RESET);
         } else {
-            System.out.println(Main.ANSI_RED + "Книги с таким id нет в списке" + Main.ANSI_RESET);
+            System.out.println(Menu.ANSI_RED + "Книги с таким id нет в списке" + Menu.ANSI_RESET);
         }
     }
 
 
-    //v1 public static void addBooksToXML(Document document) {
     public static void addBooksToXML() {
 
-       // *****************************************************************************************************************
         DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
         DocumentBuilder builder;
         try {
             builder = factory.newDocumentBuilder();
-
-            // создаем пустой объект Document, в котором будем создавать наш xml-файл
             Document document = builder.newDocument();
 
-
-            // создаем корневой элемент
+            // создает корневой элемент
             Element root = document.createElement("books");
-            // добавляем корневой элемент в объект Document
             document.appendChild(root);
-
-
 
 
             for (int i = 0; i < list.size(); i++) {
 
-            Element book = document.createElement("book");
+                Element book = document.createElement("book");
 
-            Element title = document.createElement("title");
-            title.setTextContent(list.get(i).getTitle());
+                Element title = document.createElement("title");
+                title.setTextContent(list.get(i).getTitle());
 
-            Element text = document.createElement("text");
-            text.setTextContent(list.get(i).getText());
+                Element text = document.createElement("text");
+                text.setTextContent(list.get(i).getText());
 
-            Element genre = document.createElement("genre");
-            genre.setTextContent(list.get(i).getGenre().value);
+                Element genre = document.createElement("genre");
+                genre.setTextContent(list.get(i).getGenre().value);
 
-            Element publishDate = document.createElement("publishDate");
-            publishDate.setTextContent(list.get(i).getPublishDate());
+                Element publishDate = document.createElement("publishDate");
+                publishDate.setTextContent(list.get(i).getPublishDate());
 
-            Element isbn = document.createElement("ISBN");
-            isbn.setTextContent(list.get(i).getISBN());
+                Element isbn = document.createElement("ISBN");
+                isbn.setTextContent(list.get(i).getISBN());
 
-            // Добавляем внутренние элементы книги в элемент <Book>
-            book.appendChild(title);
-            book.appendChild(text);
-            book.appendChild(genre);
-            book.appendChild(publishDate);
-            book.setAttribute("isbn", list.get(i).getISBN());
+                // Добавляет внутренние элементы книги в элемент <Book>
+                book.appendChild(title);
+                book.appendChild(text);
+                book.appendChild(genre);
+                book.appendChild(publishDate);
+                book.setAttribute("isbn", list.get(i).getISBN());
 
-//            Node root = document.getDocumentElement();
-//            // Добавляем книгу в корневой элемент
-            root.appendChild(book);
-
-            // Записываем XML в файл
-            writeDocument(document);
-        }
-
-
-
+//            // Добавляет книгу в корневой элемент
+                root.appendChild(book);
+                // Записывает XML в файл
+                writeDocument(document);
+            }
 
 
         } catch (Exception e) {
             e.printStackTrace();
         }
 
-       // *****************************************************************************************************************
-
-//        for (int i = 0; i < list.size(); i++) {
-//
-//            Element book = document.createElement("book");
-//
-//            Element title = document.createElement("title");
-//            title.setTextContent(list.get(i).getTitle());
-//
-//            Element text = document.createElement("text");
-//            text.setTextContent(list.get(i).getText());
-//
-//            Element genre = document.createElement("genre");
-//            genre.setTextContent(list.get(i).getGenre().value);
-//
-//            Element publishDate = document.createElement("publishDate");
-//            publishDate.setTextContent(list.get(i).getPublishDate());
-//
-//            Element isbn = document.createElement("ISBN");
-//            isbn.setTextContent(list.get(i).getISBN());
-//
-//            // Добавляем внутренние элементы книги в элемент <Book>
-//            book.appendChild(title);
-//            book.appendChild(text);
-//            book.appendChild(genre);
-//            book.appendChild(publishDate);
-//            book.setAttribute("isbn", list.get(i).getISBN());
-//
-//            Node root = document.getDocumentElement();
-//            // Добавляем книгу в корневой элемент
-//            root.appendChild(book);
-//
-//            // Записываем XML в файл
-//            writeDocument(document);
-//        }
 
     }
 
@@ -239,7 +188,7 @@ public class Library {
             TransformerFactory transformerFactory = TransformerFactory.newInstance();
             Transformer tr = transformerFactory.newTransformer();
             DOMSource source = new DOMSource(document);
-            FileOutputStream fos = new FileOutputStream(Main.FILE_PATH + "book.xml");
+            FileOutputStream fos = new FileOutputStream(Menu.FILE_PATH + "book.xml");
             StreamResult result = new StreamResult(fos);
             tr.transform(source, result);
         } catch (TransformerException | IOException e) {
