+++ part_1/src/com/chernova/libraryXML/book.xml
@@ -1,11 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <books>
-    <book isbn="0-0000-0000-0">
-        <title>2</title>
-        <text>2</text>
-        <genre>Romance</genre>
-        <publishDate>0000-00-00</publishDate>
-    </book>
     <book isbn="2-1234-1234-2">
         <title>Midnight Rain</title>
         <text>A former architect battles corporate zombies...</text>
@@ -15,7 +9,7 @@
     <book isbn="1-1234-1234-1">
         <title>XML Developer's Guide</title>
         <text>An in-depth look at creating applications with XML</text>
-        <genre>Science</genre>
+        <genre>Adventure</genre>
         <publishDate>2000-10-01</publishDate>
     </book>
 </books>
