--- part_1/src/com/chernova/libraryXML/oldbook.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<books>
-    <book isbn="1-1234-1234-1">
-        <title>XML Developer's Guide</title>
-        <text>An in-depth look at creating applications with XML</text>
-        <genre>Science</genre>
-        <publishDate>2000-10-01</publishDate>
-    </book>
-
-    <book isbn="2-1234-1234-2">
-        <title>Midnight Rain</title>
-        <text>A former architect battles corporate zombies...</text>
-        <genre>Fantasy</genre>
-        <publishDate>2000-12-16</publishDate>
-    </book>
-
-
-
-</books>
